Match RoomLoadTileset

+++ include/data/clipdata_types_tilemap.h
@@ -4,10 +4,10 @@
 #include "types.h"
 
 extern const u16 sClipdataBehaviorTypes_Tilemap[208];
-extern const u16 sClipdataCollisionTypes_Tilemap[208];
+extern const u8 sClipdataCollisionTypes_Tilemap[208];
 
 extern const u16 sCommonTilemap[832];
-extern const u16 sClipdataCollisionTypes_Debug[160];
+extern const u8 sClipdataCollisionTypes_Debug[160];
 extern const u16 sClipdataBehaviorTypes_Debug[160];
 
 #endif /* CLIPDATA_TYPES_TILEMAP_H */
